zero disks only for vmware and vbox

+++ scripts/zerodisk.sh
@@ -1,5 +1,11 @@
 #!/bin/bash
 
+VIRT=$(facter virtual)
+
+if [[ "$VIRT" != "vmware" ]] || [[ "$VIRT" != "virtualbox" ]]; then
+    exit 0
+fi
+
 # Zero out the free space to save space in the final image:
 dd if=/dev/zero of=/EMPTY bs=1M
 rm -f /EMPTY
